+++ lib/volt/extra_core/logger.rb
@@ -87,4 +87,3 @@ else
     end
   end
 end
-