+++ lib/volt/models/model.rb
@@ -40,7 +40,7 @@ module Volt
 
     # the id is stored in a field named _id, so we setup _id to proxy to this
     def _id
-      @attributes && @attributes[:_id]
+      __id
     end
 
     def _id=(val)
@@ -64,7 +64,7 @@ module Volt
 
       if attrs
         # Assign id first
-        id       = attrs.delete(:_id)
+        id = attrs.delete(:_id)
 
         # When doing a mass-assign, we don't save until the end.
         Model.nosave do