+++ spec/server/html_parser/view_parser_spec.rb
@@ -51,7 +51,6 @@ describe Volt::ViewParser do
                                      ]
                                    }
                                  })
-
   end
 
   it "should handle nested if's" do
@@ -117,11 +116,10 @@ describe Volt::ViewParser do
                                    'html' => "      <div class=\"main\">\n        <!-- $0 --><!-- $/0 -->\n      </div>\n",
                                    'bindings' => {
                                      0 => [
-                                       "lambda { |__p, __t, __c, __id| Volt::EachBinding.new(__p, __t, __c, __id, Proc.new { _items }, \"item\", \"main/main/main/body/__each0/__template/0\") }"
+                                       "lambda { |__p, __t, __c, __id| Volt::EachBinding.new(__p, __t, __c, __id, Proc.new { _items }, \"item\", nil, \"main/main/main/body/__each0/__template/0\") }"
                                      ]
                                    }
                                  })
-
   end
 
   it 'should parse a single attribute binding' do
@@ -201,7 +199,6 @@ describe Volt::ViewParser do
   end
 
   it 'should parse components' do
-
   end
 
   it 'should parse sections' do
@@ -221,7 +218,6 @@ describe Volt::ViewParser do
                                  'main/main/main/body' => {
                                    'html' => "\n      <p>This text goes in the body</p>\n"
                                  })
-
   end
 
   it 'should keep the html inside of a textarea if there are no bindings' do