+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 ### Added
 - Added email validator
 - each_with_index is now supported in views and the ```index``` value is no longer provided by default.
-
+- fixed bug with cookie parsing with equals in them
+- fixed bug appending existing models to a collection
 
 ## 0.8.24 - 2014-12-05
 ### Added