+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change Log
 
-## 0.8.25 - ...
+## 0.8.27 - WIP
 ### Added
 - Added email validator
-
+- each_with_index is now supported in views and the ```index``` value is no longer provided by default.
+- fixed bug with cookie parsing with equals in them
+- fixed bug appending existing models to a collection
+- refactored TemplateBinding, moved code into ViewLookupForPath (SRP)
 
 ## 0.8.24 - 2014-12-05
 ### Added